+++ src/librarian/library_private.h
@@ -82,7 +82,7 @@ typedef struct library_s {
     kh_bridgemap_t      *gbridgemap;    // global symbol bridgemap
     kh_bridgemap_t      *wbridgemap;    // weak symbol bridgemap
     kh_bridgemap_t      *lbridgemap;    // local symbol bridgemap
-    int                 dlopen;   // idx to the dlopen idx (or 0 if not dlopen)
+    size_t              dlopen;   // idx to the dlopen idx (or 0 if not dlopen)
 } library_t;
 
 // type for map elements
@@ -124,6 +124,6 @@ void removeLinkMapLib(library_t* lib);
 int FiniLibrary(library_t* lib, x64emu_t* emu);
 void Free1Library(library_t **lib, x64emu_t* emu);
 
-void RemoveDlopen(library_t** lib, int idx); // defined in wrappedlibdl.c
+void RemoveDlopen(library_t** lib, size_t idx); // defined in wrappedlibdl.c
 
 #endif //__LIBRARY_PRIVATE_H_
